iio-backend.h: Export utility function iio_read_counter_us()
This function can be used to get the value of a counter with a precision
of a microsecond. The value returned is not properly defined, and as
such does not allow to represent a specific point in time, but allows to
compute periods and delays.
