+++ iiod-responder.h
@@ -82,9 +82,6 @@ void iiod_responder_destroy(struct iiod_responder *responder);
 void iiod_responder_set_timeout(struct iiod_responder *priv,
 				unsigned int timeout_ms);
 
-/* Read the current value of the micro-second counter */
-uint64_t iiod_responder_read_counter_us(void);
-
 /* Wait until the iiod_responder stops. */
 void iiod_responder_wait_done(struct iiod_responder *responder);
