+++ include/iio/iio-backend.h
@@ -11,6 +11,7 @@
 #include <iio/iio.h>
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #define __api __iio_api
 
@@ -169,6 +170,7 @@ __api __iio_printf ssize_t
 iio_snprintf(char *buf, size_t len, const char *fmt, ...);
 __api char *iio_strdup(const char *str);
 __api size_t iio_strlcpy(char * __restrict dst, const char * __restrict src, size_t dsize);
+__api uint64_t iio_read_counter_us(void);
 
 __api struct iio_context *
 iio_create_context_from_xml(const struct iio_context_params *params,
